move methods from CommonUtils to web/utils.
